+++ test/kafka-stress/test-results.json
@@ -1,13 +1,13 @@
 {
-  "TestKafkaAsyncSinkStress/.tps": 400,
-  "TestKafkaSinkStress/.tps": 250,
+  "TestKafkaAsyncSinkStress/.tps": 250,
+  "TestKafkaSinkStress/.tps": 150,
   "TestKafkaSinkStress/N=10,messageSize=100.tps": 200,
   "TestKafkaSinkStress/N=10,messageSize=1000.tps": 150,
   "TestKafkaSinkStress/N=50000.tps": 750,
   "TestKafkaSinkStress/async=true.tps": 400,
   "TestKafkaSinkStress/messageSize=1000.tps": 300,
   "TestKafkaSinkStress/replicas=2.tps": 400,
-  "TestKafkaSourceStress/.tps": 400,
+  "TestKafkaSourceStress/.tps": 500,
   "TestKafkaSourceStress/N=10,messageSize=100.tps": 450,
   "TestKafkaSourceStress/N=10,messageSize=1000.tps": 650,
   "TestKafkaSourceStress/N=50000.tps": 3150,
